requirements.txt
tweepy>=4.14.0
nltk>=3.8.1
matplotlib>=3.7.0
wordcloud>=1.9.2
dash>=2.14.0
plotly>=5.18.0
python-dotenv>=1.0.0
